DHW target: discrete °C select dropdown instead of a number/box
A whole-degree dropdown (47–60 °C) is a single, unambiguous write per
change. The number/box stepper sent one cloud write per increment and
raced over the slow round-trip — it missed steps and HA/app disagreed.

- select.py: add WarmlinkDHWTargetSelect (R01); plain-int options avoid a
  "°C"-suffix encoding error; optimistic update for instant UI feedback.
- __init__.py: drop "number" from PLATFORMS; remove number.py.
- CHANGELOG: v71.0 now lists the DHW target as a select, not a number.

Live-tested on hardware: select 52→R01 52.0, select 55→R01 55.0.
