+++ custom_components/warmlink/__init__.py
@@ -5,7 +5,7 @@ import logging
 
 LOGGER = logging.getLogger(__name__)
 
-PLATFORMS = ["sensor", "button", "switch", "select", "number"]
+PLATFORMS = ["sensor", "button", "switch", "select"]
 
 async def async_setup(hass, config):
     """Set up the WarmLink component."""
