+++ custom_components/warmlink/select.py
@@ -43,8 +43,11 @@ VALUE_TO_LABEL = {v: k for k, v in MODE_OPTIONS.items()}
 async def async_setup_entry(hass, entry, async_add_entities):
     """Set up WarmLink select entities."""
     coordinator = hass.data[DOMAIN][entry.entry_id]
-    async_add_entities([WarmlinkModeSelect(coordinator, entry)])
-    LOGGER.info("WarmLink: Added operating mode select")
+    async_add_entities([
+        WarmlinkModeSelect(coordinator, entry),
+        WarmlinkDHWTargetSelect(coordinator, entry),
+    ])
+    LOGGER.info("WarmLink: Added operating mode + DHW target selects")
 
 
 class WarmlinkModeSelect(CoordinatorEntity, SelectEntity):
@@ -128,3 +131,106 @@ class WarmlinkModeSelect(CoordinatorEntity, SelectEntity):
         LOGGER.info("WarmLink: Mode=%s command response: %s", value, resp)
         # Refresh so the select reflects the new state from the API.
         await self.coordinator.async_request_refresh()
+
+
+# DHW target temperature as a discrete °C dropdown (writes the R01 code).
+# The device's DHW target range is R36..R37 = 47..60 °C.
+DHW_TARGET_CODE = "R01"
+DHW_MIN = 47
+DHW_MAX = 60
+
+
+class WarmlinkDHWTargetSelect(CoordinatorEntity, SelectEntity):
+    """Writable DHW target temperature as a discrete °C dropdown (R01).
+
+    A dropdown of whole-degree options is a single, unambiguous write per
+    change — unlike a stepper/box, which sends one cloud write per increment
+    and races over the slow cloud round-trip.
+    """
+
+    def __init__(self, coordinator, entry):
+        """Initialize the DHW target select."""
+        super().__init__(coordinator)
+        self._entry = entry
+        self._attr_unique_id = f"{entry.entry_id}_dhw_target_select"
+        self._attr_name = "DHW Target Temperature"
+        self._attr_icon = "mdi:thermometer-water"
+        self._attr_options = [str(t) for t in range(DHW_MIN, DHW_MAX + 1)]
+
+    @property
+    def device_info(self) -> DeviceInfo:
+        """Return device info (matches the sensors/switch/select device)."""
+        device_name = "WarmLink"
+        device_model = "Heat Pump"
+
+        if self.coordinator.device_info:
+            nick = self.coordinator.device_info.get("device_nick_name")
+            cust_model = self.coordinator.device_info.get("cust_model")
+
+            if nick and nick.strip():
+                device_name = nick
+            elif cust_model and cust_model.strip():
+                device_name = cust_model
+
+            if cust_model and cust_model.strip():
+                device_model = cust_model
+
+        return DeviceInfo(
+            identifiers={(DOMAIN, self._entry.entry_id)},
+            name=device_name,
+            manufacturer="WarmLink",
+            model=device_model,
+        )
+
+    def _r01_value(self):
+        """Return the raw R01 value, or None if unavailable."""
+        if self.coordinator.data:
+            for item in self.coordinator.data:
+                if item.get("code") == DHW_TARGET_CODE:
+                    return item.get("value")
+        return None
+
+    @property
+    def available(self) -> bool:
+        """Available as long as the coordinator has data."""
+        return bool(self.coordinator.data) and self.coordinator.last_update_success
+
+    @property
+    def current_option(self):
+        """Return the current DHW target as a dropdown option (rounded to °C)."""
+        value = self._r01_value()
+        if value in (None, "", "null"):
+            return None
+        try:
+            opt = str(int(round(float(value))))
+        except (TypeError, ValueError):
+            return None
+        return opt if opt in self._attr_options else None
+
+    async def async_select_option(self, option: str) -> None:
+        """Write the chosen DHW target temperature to the device."""
+        try:
+            target = int(option)
+        except (ValueError, TypeError):
+            LOGGER.error("WarmLink: Unrecognised DHW target option %s", option)
+            return
+
+        device_code = None
+        if self.coordinator.device_info:
+            device_code = self.coordinator.device_info.get("device_code")
+        if not device_code:
+            LOGGER.error("WarmLink: No device_code available, cannot set DHW target")
+            return
+
+        out = f"{target}.0"
+        LOGGER.info("WarmLink: Requesting DHW target R01=%s (select)", out)
+        resp = await self.coordinator.api.set_value(device_code, DHW_TARGET_CODE, out)
+        LOGGER.info("WarmLink: R01=%s select response: %s", out, resp)
+        # Optimistically reflect the new value so the dropdown updates instantly;
+        # the next scheduled poll reconciles (and corrects if the device clamps).
+        if self.coordinator.data:
+            for item in self.coordinator.data:
+                if item.get("code") == DHW_TARGET_CODE:
+                    item["value"] = out
+                    break
+            self.coordinator.async_set_updated_data(self.coordinator.data)
